Use non-breaking space after title abbreviation
LaTex interprets a space after a period as the end of a sentence, and adds
extra space after it accordingly in the output.  Abbreviations are not ends of
sentences, though, and should not have this extra space.  Using the non-
breaking space character `~` instead of a regular space corrects this.

Also removed comment about how LaTeX paragraphs work, since it was rather
overly-verbose--I expect that anyone using this template will already know that
rather basic information.